backend/requirements.txt
PyJWT==1.5.3
tornado==4.5.2
redis==4.1.2
aiopg==1.3.3
sqlalchemy==1.1.15
sqlalchemy_utils==0.36.8
psycopg2==2.9.3
alembic==0.9.6
# apispec==3.1.0
pytest==3.2.5
webargs==5.5.2
passlib==1.7.1
hashids==1.2.0
python-magic==0.4.15
requests==2.27.1
celery==5.2.3
pyssdb==0.4.1
boto3==1.20.49
apispec-webframeworks==0.5.2
PyYAML==6.0
apispec==5.1.1
factory-boy==3.2.1
